clean up and separation
